requirements.txt
fastapi
uvicorn[standard]
openai
supabase-python
python-dotenv
pydantic
dspy
